refactor: move ai chat billing card into chat feature
