+++ client/src/features/chat/components/ai-chat-billing-card.tsx
